+++ itolib/Behaviours/Events/ToggleEvent.cs
@@ -1,0 +1,40 @@
+using UnityEngine;
+using UnityEngine.Events;
+
+namespace itolib.Behaviours.Events
+{
+    /// <summary>
+    ///     TODO.
+    /// </summary>
+    public class ToggleEvent : MonoBehaviour // TODO: Networked?
+    {
+        /// <summary>
+        ///     TODO.
+        /// </summary>
+        [Header("Toggle Event")]
+        [Tooltip("")]
+        public UnityEvent? onEnable;
+
+        /// <summary>
+        ///     TODO.
+        /// </summary>
+        [Tooltip("")]
+        public UnityEvent? onDisable;
+
+        /// <summary>
+        ///     TODO.
+        /// </summary>
+        public void OnEnable()
+        {
+            onEnable?.Invoke();
+        }
+
+        /// <summary>
+        ///     TODO.
+        /// </summary>
+        public void OnDisable()
+        {
+            onDisable?.Invoke();
+        }
+    }
+}